+++ pytsite/file/_model.py
@@ -178,7 +178,7 @@ class AbstractImage(AbstractFile):
 
         return '<span class="{}" data-url="{}" data-alt="{}" data-aspect-ratio="{}" ' \
                'data-width="{}" data-height="{}" data-enlarge="{}"></span>' \
-            .format(css.strip(), self.get_url(), alt, aspect_ratio, self.width, self.height, enlarge)
+            .format(css.strip(), self.url, alt, aspect_ratio, self.width, self.height, enlarge)
 
     def as_jsonable(self, **kwargs) -> dict:
         r = super().as_jsonable(**kwargs)